backend/requirements.txt
arrow==1.2.1
asgiref==3.4.1
backports.zoneinfo==0.2.1
blessed==1.19.0
Django==4.0.1
django-cors-headers==3.11.0
django-extensions==3.1.5
django-picklefield==3.0.1
django-probes==1.6.0
django-q==1.3.9
django-redis==5.2.0
django-rest-framework==0.1.0
djangorestframework==3.13.1
psycopg2==2.9.3
python-dateutil==2.8.2
pytz==2021.3
redis==3.5.3
six==1.16.0
sqlparse==0.4.2
uWSGI==2.0.20
wcwidth==0.2.5
